pyproject.toml
[tool.poetry]
name = "repl_nix_mi-haels-badiebadumdumdumdumdumdumdumdumdumdumdum"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.21.0"
requests = "^2.28.1"
aiohttp = "^3.8.3"
discord = "^2.0.0"
ffmpeg = "^1.4"
mutagen = "^1.46.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
